api/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
SQLAlchemy==1.4.25
passlib==1.7.4
PyJWT==2.1.0
